src/requirements.txt
asgiref==3.7.2
Django==4.2.6
packaging==23.2
Pillow==10.1.0
pytesseract==0.3.10
sqlparse==0.4.4
tzdata==2023.3
